Define toybox_recovery
"toybox" module currently sets `recovery_available: true` to install the
module and the symlinks to the recovery partition in make. However, this
is not applicable in Soong because `recovery_available` is used to allow
the reverse dependencies of this module installed in recovery partition
to be depend on this module, and is not used for installation of the
module itself to the recovery partition.

As part of the effort to migrate the installation rule of the recovery
partition from make to Soong, this change introduces "toybox_recovery"
module to install the artifacts to the recovery partition, similar to
"toybox_vendor".

Test: m soong_generated_recovery_filesystem_test
Bug: 381888358
Change-Id: I4d4761fc0bf558f4f89874f39b7281491a72507c